+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mybenzar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 17:51:54 by mybenzar          #+#    #+#             */
-/*   Updated: 2018/11/13 19:38:59 by mybenzar         ###   ########.fr       */
+/*   Updated: 2018/11/19 16:48:49 by mybenzar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
 		return (0);
 	while ((unsigned char)s1[i] != '\0'
-			&& (unsigned char)s2[i] != '\0' && i < (int)n)
+			&& (unsigned char)s2[i] != '\0' && i < n)
 	{
 		if ((unsigned char)s1[i] != (unsigned char)s2[i])
 			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
@@ -28,7 +28,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	}
 	if ((((unsigned char)s1[i] == '\0' && (unsigned char)s2[i] != '\0')
 				|| ((unsigned char)s2[i] == '\0'
-					&& (unsigned char)s1[i] != '\0')) && (i < (int)n))
+					&& (unsigned char)s1[i] != '\0')) && (i < n))
 		return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 	return (0);
 }
