+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: mybenzar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 11:45:08 by mybenzar          #+#    #+#             */
-/*   Updated: 2018/11/09 13:46:45 by mybenzar         ###   ########.fr       */
+/*   Updated: 2018/11/19 16:31:35 by mybenzar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
-	while (src[i] != '\0' && i < (int)len)
+	while (src[i] != '\0' && i < len)
 	{
 		dst[i] = src[i];
 		i++;
 	}
-	while (i < (int)len)
+	while (i < len)
 	{
 		dst[i] = '\0';
 		i++;
